% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt5reti.R
\name{copy_rates_range}
\alias{copy_rates_range}
\title{copy_rates_range}
\usage{
copy_rates_range(symbol, timeframe, date_from, date_to)
}
\arguments{
\item{symbol}{character}

\item{timeframe}{mt5 timeframe}

\item{date_from}{int}

\item{date_to}{int}
}
\value{
\itemize{
\item 
}
}
\description{
copy_rates_range
}
