% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt5reti.R
\name{market_book_release}
\alias{market_book_release}
\title{market_book_release}
\usage{
market_book_release(symbol)
}
\arguments{
\item{symbol}{character}
}
\value{
\itemize{
\item 
}
}
\description{
market_book_release
}
