% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.AccountInfo}
\alias{MT5.AccountInfo}
\title{Account information}
\usage{
MT5.AccountInfo()
}
\value{
\emph{Data.frame} \eqn{[1x11]} is returned, with follow informations:
\itemize{
  \item Company \code{{character}}: company name.
  \item Acc_server \code{{character}}: server name that your MT5 is connected to.
  \item Acc_number \code{{int}}: is the account login as well.
  \item Balance \code{{numeric}}: same used in MT5 Trade tab.
  \item Equity  \code{{numeric}}: is the sum of Balance and current Profit.
  \item Profit \code{{numeric}}: sum of all open profit/loss positions.
  \item Credit \code{{numeric}}: same used in MT5 Trade tab.
  \item Margin \code{{numeric}}: is the amount of money necessary to cover your possible losses during margin trading.
  \item Margin_Free \code{{numeric}}: is the amount availabe to open next trades. Free margin equals equity minus margin.
  \item {Account_type \code{{int}}:
         \itemize{
                 \item{0: ACCOUNT_TRADE_MODE_REAL (real account)}
                 \item{1: ACCOUNT_TRADE_MODE_DEMO (demo account)}
                 \item{2: ACCOUNT_TRADE_MODE_CONTEST}
                 }
         }
  \item {Margin_type \code{{int}}:
         \itemize{
                 \item{0: ACCOUNT_MARGIN_MODE_RETAIL_NETTING (Netting)}
                 \item{1: ACCOUNT_MARGIN_MODE_EXCHANGE (Exchange)}
                 \item{2: ACCOUNT_MARGIN_MODE_RETAIL_HEDGING (Hedging)}
                 }
         }
}
}
\description{
This function is useful for obtaining MT5's account information.
}
\references{
\url{https://www.mql5.com/en/docs/constants/environment_state/accountinformation}

\url{https://hercules.finance/faq/what-is-balance-equity-margin-and-free-margin-and-margin-call/}
}
\seealso{
\link{\code{MT5.ShowPositions()}}, \link{\code{MT5.ShowOrders()}}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
