% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.AllSymbols}
\alias{MT5.AllSymbols}
\title{Fetch all symbols}
\usage{
MT5.AllSymbols()
}
\value{
Returns \code{character()} of all symbols available in MT5.
}
\description{
Fetch all symbols, even those not on Marketwatch.

This function may take time.
}
\examples{
\dontrun{

MT5.AllSymbols()

}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
