% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.BidAskBook}
\alias{MT5.BidAskBook}
\title{Order book}
\usage{
MT5.BidAskBook(sSymbol, iBidAsk = 0, iBookDepth = 5)
}
\arguments{
\item{sSymbol}{character; target symbol}

\item{iBidAsk}{int; if \code{iBidAsk = 0} Bid is given, \code{iBidAsk = 1} Ask is given (default \code{0}).}

\item{iBookDepth}{int; book depth (default \code{5})}
}
\value{
\emph{Data.frame} \eqn{[nx2]} is returned, with follow informations:
\itemize{
\item fPrice \code{{numeric}}: price
\item Volume \code{{int}}: bid/ask grouped by price
}
}
\description{
This function is for obtaining Bid or Ask order book with specified book depth.
}
\details{
(Stocks) For open and closing auctions this function cannot reflect the book properly.
}
\references{
\url{https://www.investopedia.com/terms/o/order-book.asp}
}
\seealso{
\code{\link{MT5.BidAskSpread()}}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
