% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.BidAskSpread}
\alias{MT5.BidAskSpread}
\title{Order book Spread}
\usage{
MT5.BidAskSpread(sSymbol)
}
\arguments{
\item{sSymbol}{character; target symbol.}
}
\value{
\emph{Data.frame} \eqn{[1x4]} is returned, with follow informations:
\itemize{
\item sSymbol \code{{character}}: symbol used.
\item fBid \code{{numeric}}: higher bid.
\item fAsk \code{{numeric}}: lower ask.
\item Spread \code{{numeric}}: difference between lower ask and higher bid.
}
}
\description{
This function is for obtaining the higher bid, the lower ask and the spread between them.
}
\details{
For same result, use \code{MT5.BidAskBook()} and calculate the difference between the best bid and ask.

For Forex brokers this function will return an empty \code{data.frame}.
}
\references{
\url{https://www.investopedia.com/terms/o/order-book.asp}
}
\seealso{
\code{\link[=MT5.BidAskBook]{MT5.BidAskBook()}}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
