% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.CheckVersion}
\alias{MT5.CheckVersion}
\title{Check mt5R version in R and MT5}
\usage{
MT5.CheckVersion()
}
\value{
Returns \code{TRUE} if MT5 and R have the same version of mt5R, \code{FALSE} otherwise.

A warning message will be displayed in case of \code{FALSE} return.
}
\description{
This function is useful to verify both versions of mt5R. It is advisable to use the same version in R and MT5.

In case of \code{FALSE} return consider to update your mt5R package or mt5R EA, check for warning messages in this case.
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
