% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.Connect}
\alias{MT5.Connect}
\title{Socket connection with MT5}
\usage{
MT5.Connect(sReq, iPort = 23456, bMsg = FALSE, timeout = getOption("timeout"))
}
\arguments{
\item{sReq}{character; requisition sent to MT5.}

\item{iPort}{int; port used to create socket connection. (default \code{23456})}

\item{bMsg}{bool; if \code{TRUE} it will print all return messages from MT5. (default \code{FALSE})}

\item{timeout}{int; timeout used in \code{\link[base:connections]{base::socketConnection()}}. (default \code{60})}
}
\value{
Returns \code{character}, it will be entirely dependent of the function of the other side in the mt5R package running in MT5. Error due connection will be returned \code{NULL}.
}
\description{
The main function of mt5R, its make the connection by socket with MT5. Used in all mt5R functions.

This function was only viable by JC's package (Socket library for MT4 and MT5). See reference.
}
\details{
It uses <socket-library-mt4-mt5.mqh> library.

This function still has some experiments concepts, especially with delayed functions that MT5 requires time to run.
}
\references{
\url{https://www.mql5.com/en/blogs/post/706665}
}
\seealso{
\code{\link[=MT5.ChangeDoorSocket]{MT5.ChangeDoorSocket()}}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
