% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.DeleteOrder}
\alias{MT5.DeleteOrder}
\title{Delete pending order}
\usage{
MT5.DeleteOrder(iTickets)
}
\arguments{
\item{int}{int(); vector of tickets orders to delete.}
}
\value{
Returns \code{TRUE} if order was successful deleted, \code{FALSE} otherwise.
}
\description{
Send a request to delete orders in MT5.
}
\details{
For iTickers users can use \code{\link[=MT5.ShowOrders]{MT5.ShowOrders()}} to obtain it.

(STOCKS ONLY) WARNING - Poor written MT5's servers

I had some problems running this function in some brokers and doesn't in others.
All brokers that are winged by XP Investimentos (BR broker), the MT5's server try to close another nonexistent order (order #0)
The same error don't occurs in Modal (BR broker) server. In all cases, the order is deleted anyway.
Use this function very carefully if you notice that the MT5's server broker was poor written.
I already tried to fix into MQL code in every way. All forms to remove the order result in the same outcome.
For who is experiencing this problem, take note that the mt5R plataform takes approximately 3 minutes to recover itself.
Poor fix: the trader can use \code{\link[=MT5.ModifyOrder]{MT5.ModifyOrder()}} to change fPrice to unviable levels and waits to order expire.
}
\examples{
\dontrun{
# Delete pending order 832136429
MT5.DeleteOrder(832136429)

}
}
\seealso{
\code{\link[=MT5.SingleOrder]{MT5.SingleOrder()}}, \code{\link[=MT5.MultipleOrders]{MT5.MultipleOrders()}}, \code{\link[=MT5.ShowPositions]{MT5.ShowPositions()}}, \code{\link[=MT5.ShowOrders]{MT5.ShowOrders()}}, \code{\link[=MT5.ModifyOrder]{MT5.ModifyOrder()}}, \code{\link[=MT5.ClosePosition]{MT5.ClosePosition()}}, \code{\link[=MT5.ModifyPosition]{MT5.ModifyPosition()}}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
