% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.DrawHorizontalLine}
\alias{MT5.DrawHorizontalLine}
\title{Draw horizontal line}
\usage{
MT5.DrawHorizontalLine(
  sSymbol,
  fPrice,
  iWidth = 1,
  iColor = 1,
  sName = "",
  iTF = 0
)
}
\arguments{
\item{sSymbol}{character; target symbol.}

\item{fPrice}{numeric; at which price will be draw the line.}

\item{iWidth}{int; which width will be draw the line (default \code{1}).}

\item{iColor}{int; See details (default \code{1}: \code{clrGreen}).}

\item{sName}{character; unique name if you want further modifications in lines parameters.}

\item{iTF}{int; its always in minutes. If \code{iTF = 0} it will be draw on all time frames (default \code{0}). See details.}
}
\value{
Returns \code{TRUE} if the line was drawn, \code{FALSE} otherwise.
}
\description{
Draw horizontal line into MT5 chart(s).
}
\details{
Supported colors (\code{iColor}). See url below for even more details.
\itemize{
\item{1: clrGreen (GREEN)}
\item{2: clrBlue (BLUE)}
\item{3: clrRed (RED)}
\item{4: clrNavy	}
\item{5: clrPurple	}
\item{6: clrIndigo	}
\item{7: clrOliveDrab	}
\item{8: clrDarkSlateBlue	}
\item{9: clrLawnGreen	}
\item{10: clrOrangeRed	}
\item{11: clrGold	}
\item{12: clrYellow	}
\item{13: clrAqua	}
\item{14: clrMagenta	}
\item{15: clrLightSlateGray	}
\item{16: clrPaleVioletRed	}
\item{17: clrHotPink	}
\item{18: clrKhaki	}
\item{19: clrSilver	}
\item{20: clrLightGray	}
\item{21: clrKhaki	}
\item{22: clrBeige	}
}
\url{https://www.mql5.com/en/docs/constants/objectconstants/webcolors}


Supported time frames (\code{iTF}). See url below for even more details.
\itemize{
\item{0: all	(it will be draw on all time frames)}
\item{1: \code{PERIOD_M1}	}
\item{2: \code{PERIOD_M2}	}
\item{5: \code{PERIOD_M5}	}
\item{15: \code{PERIOD_M15}	}
\item{30: \code{PERIOD_M30}	}
\item{60: \code{PERIOD_H1}	}
\item{120: \code{PERIOD_H2}	}
\item{240: \code{PERIOD_H4}	}
\item{480: \code{PERIOD_H8}	}
\item{1440: \code{PERIOD_D1}	}
\item{7200: \code{PERIOD_W1}	}
\item{216000: \code{PERIOD_MN1}	}
}
\url{https://www.mql5.com/en/docs/constants/chartconstants/enum_timeframes}
}
\examples{
\dontrun{
# Red line in all EURUSD charts at 1.1772
MT5.DrawHorizontalLine("EURUSD", 1.1772, iColor = 3)

}
}
\seealso{
\link{\code{MT5.RemoveAllChartsObjects()}},
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
