% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.FileWorkingFolder}
\alias{MT5.FileWorkingFolder}
\title{MT5 File folder}
\usage{
MT5.FileWorkingFolder()
}
\value{
Returns file path of MT5 folder.
}
\description{
Used in functions that need to use files generated from MT5.
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
