% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.ListOptions}
\alias{MT5.ListOptions}
\title{List options of an underlying asset}
\usage{
MT5.ListOptions(
  sUnderlyingAsset,
  bUseDescriptionStrike = F,
  bAutomaticHelpUnderlying = T
)
}
\arguments{
\item{sUnderlyingAsset}{character(); target underlying asset (e.g. PETR4 = PETR; BBDC4 = BBDC).}

\item{bUseDescriptionStrike}{bool; use strike given in symbol description. Some brokers do not adjust the prices following the dividends. Totally optional (experimental). (default \code{FALSE}).}

\item{bAutomaticHelpUnderlying}{bool; it will try to help the user to only send the underlying asset to mt5R. It will convert "PETR4" to "PETR" automatically (default \code{TRUE}).}
}
\value{
Returns \emph{Data.frame} \eqn{[nx6]}, with follow informations:
\itemize{
  \item sOpt \code{{character}}: options symbol.
  \item PN_ON \code{{int}}: 0 for preferential stocks, 1 otherwise.
  \item Type \code{{factor}}: options type. "CALL" for call options, "PUT" for put options.
  \item Mod \code{{factor}}: option style. "AMERICAN" or "EUROPEAN". See references.
  \item Strike \code{{numeric}}: options strike.
  \item Expiration \code{{Date}}: options expiration.
  }
}
\description{
Function to load all available options in MT5 of an underlying asset.

Only works in stock brokers.
}
\examples{
\dontrun{
## Get all available options of PETR4 and PETR3 (Petroleo Brasileiro S.A. - Petrobras PBR:NYSE)
MT5.ListOptions("PETR")

## Other examples
MT5.ListOptions("VALE")
MT5.ListOptions("USIM")

}
}
\references{
\url{https://en.wikipedia.org/wiki/Option_style}
}
\seealso{
\code{\link{MT5.SymbolType()}}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
