% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.MarketIsOpen}
\alias{MT5.MarketIsOpen}
\title{Check if market is open}
\usage{
MT5.MarketIsOpen()
}
\value{
Returns \code{TRUE} if market is open, \code{FALSE} otherwise.
}
\description{
\strong{Warning!} This function is not entirely reliable. It uses the last hour of trading of all marketwatch's symbols to check if the market is open. To actually know when the market is really open, the trader must know the market and rely on his/her knowledge.

If trader's computer \emph{lost connection} and all symbols turns outdated this function will return \code{FALSE} by its construction. It's strongly recommended to use support function, as check if internet is working and check for weekends/holidays, to not stuck with False Negatives.

Detecting this type of status is quite challenging in MQL. \strong{This function takes \code{1} second to return the result.}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
