% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.Marketwatch}
\alias{MT5.Marketwatch}
\title{Symbols in MT5's marketwatch}
\usage{
MT5.Marketwatch()
}
\value{
Returns \emph{character()} of all symbols in MT5's marketwatch.
}
\description{
List all symbols in MT5's marketwatch.
}
\seealso{
\code{\link{MT5.MarketwatchAdd()}}, \code{\link{MT5.MarketwatchRemove()}}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
