% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.MarketwatchAdd}
\alias{MT5.MarketwatchAdd}
\title{Add symbols in MT5's marketwatch}
\usage{
MT5.MarketwatchAdd(sSymbols)
}
\arguments{
\item{sSymbol}{character(); vector of target symbols to add.}
}
\value{
Returns \emph{logical()}. \code{TRUE} if symbol was successfully added, \code{FALSE} otherwise.
}
\description{
Add symbols in MT5's marketwatch.
}
\examples{
\dontrun{
## Add XAUUSD and XAGUSD
MT5.MarketwatchAdd(c("XAUUSD","XAGUSD"))

}
}
\seealso{
\code{\link[=MT5.Marketwatch]{MT5.Marketwatch()}}, \code{\link[=MT5.MarketwatchRemove]{MT5.MarketwatchRemove()}}, \code{\link[=MT5.SymbolInMarketwatch]{MT5.SymbolInMarketwatch()}}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
