% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.MarketwatchRemove}
\alias{MT5.MarketwatchRemove}
\title{Remove symbols in MT5's marketwatch}
\usage{
MT5.MarketwatchRemove(sSymbols)
}
\arguments{
\item{sSymbol}{character(); vector of target symbols to remove.}
}
\value{
Returns \emph{logical()}. \code{TRUE} if symbol was successfully removed, \code{FALSE} otherwise.
}
\description{
Remove symbols in MT5's marketwatch.
}
\examples{
\dontrun{
## Remove XAUUSD and XAGUSD
MT5.MarketwatchAdd(c("XAUUSD","XAGUSD"))

}
}
\seealso{
\code{\link{MT5.Marketwatch()}}, \code{\link{MT5.MarketwatchAdd()}}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
