% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.ModifyOrder}
\alias{MT5.ModifyOrder}
\title{Modify order}
\usage{
MT5.ModifyOrder(iTicket = 0, fPrice = 0, fStop = 0, fGain = 0, ...)
}
\arguments{
\item{iTicket}{integer; order's ticket.}

\item{fPrice}{numeric; target price. To no change use \code{fPrice = 0} or omit.}

\item{fStop}{numeric; target stop loss. To remove stop loss use \code{fStop = -1}. To no change use \code{fStop = 0} or omit.}

\item{fGain}{numeric; target stop gain. To remove stop gain use \code{fGain = -1}. To no change use \code{fGain = 0} or omit.}

\item{...}{data.frame; using \code{data.frame} for several orders at once. \strong{Columns name should be the same of the arguments above}. See examples.}
}
\value{
Returns \code{TRUE} if order was successful modified, no changes and otherwise will return \code{FALSE}.

If used \code{data.frame} it will return \code{logical} vector.
}
\description{
Modification of characteristics of pending orders. To modify an previously opened position use \code{MT5.ModifyPosition()} instead.

It can be used targeting only one or several orders at once. See details.

Use \code{MT5.ShowOrders()} or look at "Trade" tab, to fetch \code{iTickets}.
}
\details{
It need to choose a \code{data.frame} or just one order at time to modify. Not both. An attempt using both a warning will pop up and \code{data.frame} will be used.
}
\examples{
\dontrun{
## Below examples of an existing buy limit order (684701334)

MT5.ModifyOrder(iTicket = 684701334, fPrice = 25.00) ## Modifying the trigger price to open a limit order
MT5.ModifyOrder(iTicket = 684701334, fStop = 24.00) ## A single modification
MT5.ModifyOrder(iTicket = 684701334, fStop = 24.00, fGain = 26.00) ## You can modify both at same time
MT5.ModifyOrder(iTicket = 684701334, fPrice = 25.00, fStop = 24.00, fGain = 26.00) ## Or everything at same time
MT5.ModifyOrder(684701334, fStop = -1) ## Remove stop, omitting iTicket

## The trader can use data.frame for multiple modifications in different orders. Columns name should be the same
## of the arguments of a single modification (check used arguments above). See example below

df_modify <- data.frame(iTicket = 687356562, fPrice = 21, fStop = 20)
df_modify <- rbind(df_modify, data.frame(iTicket = 687357112, fPrice = 18, fStop = 17))
MT5.ModifyOrder(... = df_modify) ## Modify two orders at same time

}
}
\seealso{
\code{\link[=MT5.SingleOrder]{MT5.SingleOrder()}}, \code{\link[=MT5.MultipleOrders]{MT5.MultipleOrders()}}, \code{\link[=MT5.ShowOrders]{MT5.ShowOrders()}}, \code{\link[=MT5.ShowPositions]{MT5.ShowPositions()}}, \code{\link[=MT5.ClosePosition]{MT5.ClosePosition()}}, \code{\link[=MT5.DeleteOrder]{MT5.DeleteOrder()}}, \code{\link[=MT5.ModifyPosition]{MT5.ModifyPosition()}}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
