% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.MultipleOrders}
\alias{MT5.MultipleOrders}
\title{Send multiple orders}
\usage{
MT5.MultipleOrders(df_orders)
}
\arguments{
\item{df_orders}{data.frame; see details.}
}
\value{
Returns \code{{logical}} vector. \code{TRUE} if order was successful accepted, otherwise will return \code{FALSE}.
}
\description{
Send multiple orders using \code{data.frame}. All columns need to be filled. See details.

Some Forex brokers take \code{200ms} to accept orders into market, so take this time on mind. Always check MT5's "Experts" tab.
}
\details{
Take note that buy orders are rejected when order's stop loss are higher than order's price, and vice-versa for sell orders. This is a sanity check made by MT5, not by mt5R. Always check MT5's "Experts" tab for more details when an order is rejected.

\code{df_orders {data.frame}} \eqn{[nx7]} columns should be on the same order and type of \link{MT5.SingleOrder} arguments.
\itemize{
\item sSymbol \code{{character}}: symbol to send the order.
\item {iCmd \code{{int}}:
\itemize{
\item{0: ORDER_TYPE_BUY_LIMIT}
\item{1: ORDER_TYPE_SELL_LIMIT}
}
}
\item fVol \code{{numeric}}: volume order.
\item fPrice \code{{numeric}}: where the order will be positioned. Use \code{fPrice = 0} for market order.
\item fStop \code{{numeric}}: if applicable use any value > 0.
\item fGain \code{{numeric}}: if applicable use any value > 0.
\item {iFillType \code{{int}}:
\itemize{
\item{0: ORDER_FILLING_RETURN (default)}
\item{1: ORDER_FILLING_FOK (market order)}
\item{2: ORDER_FILLING_IOC}
}
}
}

Using \code{fPrice = 0} for Market Order, automatically \code{iFillType} will be set to \code{1}.
}
\examples{
\dontrun{
## Two LONG positions, PETR4 with volumn of 100 and price of 22.0 and VALE5 with volumn of 500 and price of 65.
Orders <- data.frame(sSymbol = "PETR4", iCmd = 0, fVol = 100, fPrice = 22, fStop = 0, fGain = 0, iFillType = 0)
Orders <- rbind(Orders, data.frame(sSymbol = "ABEV3", iCmd = 0, fVol = 200, fPrice = 14.3, fStop = 0, fGain = 0, iFillType = 0))
MT5.MultipleOrders(Orders)

## Forex order. First one is a market order (fPrice = 0)
Orders <- data.frame(sSymbol = "EURUSD", iCmd = 0, fVol = 0.01, fPrice = 0, fStop = 0, fGain = 0, iFillType = 1)
Orders <- rbind(Orders, data.frame(sSymbol = "GBPUSD", iCmd = 0, fVol = 0.01, fPrice = 1.3276, fStop = 0, fGain = 0, iFillType = 0))
MT5.MultipleOrders(Orders)

}
}
\references{
\url{https://www.mql5.com/en/docs/constants/tradingconstants/orderproperties}

\url{https://www.investopedia.com/terms/m/marketorder.asp}
}
\seealso{
\code{\link[=MT5.SingleOrder]{MT5.SingleOrder()}}, \code{\link[=MT5.ShowOrders]{MT5.ShowOrders()}}, \code{\link[=MT5.ShowPositions]{MT5.ShowPositions()}}, \code{\link[=MT5.ClosePosition]{MT5.ClosePosition()}}, \code{\link[=MT5.DeleteOrder]{MT5.DeleteOrder()}}, \code{\link[=MT5.ModifyOrder]{MT5.ModifyOrder()}}, \code{\link[=MT5.ModifyPosition]{MT5.ModifyPosition()}}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
