% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.Ping}
\alias{MT5.Ping}
\title{Ping MT5 platform}
\usage{
MT5.Ping(bShowTimeElapsed = F)
}
\arguments{
\item{bShowTimeElapsed}{bool; it will print how long the response takes (default \code{FALSE}). See details.}
}
\value{
Returns \code{TRUE} if mt5R is available in MT5, \code{FALSE} otherwise.
}
\description{
Signal sent to MT5 that requests a response to check if the mt5R in MT5 is available.

It can be used to measure how long the response takes.
}
\details{
For \code{bShowTime = TRUE}, is pretty common zero results. Check also ?proc.time
}
\seealso{
\code{\link{MT5.ChangeDoorSocket()}}, \code{\link{MT5.Connect()}}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
