% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.Quick_GetSymbol}
\alias{MT5.Quick_GetSymbol}
\title{Load data quickly from MT5}
\usage{
MT5.Quick_GetSymbol(sSymbol, iTF, iRows = 5, xts = FALSE)
}
\arguments{
\item{sSymbol}{character; target symbol.}

\item{iTF}{int; target time frame. See details.}

\item{xts}{bool; if \code{xts = TRUE} the function will return a xts instead. See details.}

\item{iDepth}{int; how many rows. It's start from last. (default: \code{5})}
}
\value{
Always returns \code{OHCLV} format. Date will be exhibited as it \code{xts} parameter is defined.

\describe{
\item{\code{xts = FALSE} (default)}{Returns \eqn{[nx10]} \code{{data.frame} {Year, Month, Day, Hour, Minute, Open, High, Low, Close, Volume}}}
\item{\code{xts = TRUE}}{Returns \eqn{[nx5]}  \code{{xts} {Open, High, Low, Close, Volume}}}
}
}
\description{
Function to load \code{sSymbol} in \code{Data.frame} or \code{xts} if specified. All the data come from socket. It's pretty quickly function to obtain data.

Data is loaded silently and will take longer as defined depth is.
}
\details{
It should be used only for tiny table sizes. For big size tables consider to use \code{MT5.GetSymbol()} instead.

Supported time frames (\code{iTF}). See references for even more details.
\itemize{
\item{1: \code{PERIOD_M1}}
\item{2: \code{PERIOD_M2}}
\item{5: \code{PERIOD_M5}}
\item{15: \code{PERIOD_M15}}
\item{30: \code{PERIOD_M30}}
\item{60: \code{PERIOD_H1}}
\item{120: \code{PERIOD_H2}}
\item{240: \code{PERIOD_H4}}
\item{480: \code{PERIOD_H8}}
\item{1440: \code{PERIOD_D1}}
\item{7200: \code{PERIOD_W1}}
\item{216000: \code{PERIOD_MN1}	}
}
}
\examples{
\dontrun{

MT5.Quick_GetSymbol("EURUSD", 5, 3)

## Returns
##   Year Month Day Hour Minute    Open    High     Low   Close Volume
## 1 2020    12  11    1      5 1.21411 1.21414 1.21405 1.21409     39
## 2 2020    12  11    1     10 1.21408 1.21419 1.21404 1.21418     56
## 3 2020    12  11    1     15 1.21421 1.21421 1.21405 1.21408     63

MT5.Quick_GetSymbol("EURUSD", 5, 3, xts = T)

## Returns
##                        Open    High     Low   Close Volume
## 2020-12-11 01:05:00 1.21411 1.21414 1.21405 1.21409     39
## 2020-12-11 01:10:00 1.21408 1.21419 1.21404 1.21418     56
## 2020-12-11 01:15:00 1.21421 1.21421 1.21405 1.21415     82

}
}
\references{
\url{https://en.wikipedia.org/wiki/Open-high-low-close_chart}

\url{https://www.mql5.com/en/docs/constants/chartconstants/enum_timeframes}
}
\seealso{
\code{\link{MT5.GetSymbol()}}, \code{\link{MT5.QuickLastRow_GetSymbol()}}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
