% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.RemoveAllChartsObjects}
\alias{MT5.RemoveAllChartsObjects}
\title{Remove all drawn objects}
\usage{
MT5.RemoveAllChartsObjects(sSymbol = "0", iTF = 0)
}
\arguments{
\item{sSymbol}{character(); you can specify which chart symbol. If not declared it will consider all symbols. See details.}

\item{iTF}{int; you can specify which charts time frames. If not declared it will consider all time frames. See details.}
}
\value{
Returns \code{TRUE} if succeed, \code{FALSE} otherwise.
}
\description{
Remove all drawn  objects in specifics MT5 charts.
}
\details{
\strong{If arguments are not declared} it will remove \strong{all} drawn objects of \strong{all} charts.

Supported time frames (\code{iTF}). See references for even more details.
\itemize{
\item{1: \code{PERIOD_M1}	}
\item{2: \code{PERIOD_M2}	}
\item{5: \code{PERIOD_M5}	}
\item{15: \code{PERIOD_M15}	}
\item{30: \code{PERIOD_M30}	}
\item{60: \code{PERIOD_H1}	}
\item{120: \code{PERIOD_H2}	}
\item{240: \code{PERIOD_H4}	}
\item{480: \code{PERIOD_H8}	}
\item{1440: \code{PERIOD_D1}	}
\item{7200: \code{PERIOD_W1}	}
\item{216000: \code{PERIOD_MN1}	}
}
}
\examples{
\dontrun{
## Remove all drawn objects of all GBPUSD charts
MT5.RemoveAllChartsObjects("GBPUSD")

## Remove all drawn objects of all 60 minutes (PERIOD_H1) charts
MT5.RemoveAllChartsObjects(iTF = 60)

}
}
\references{
\url{https://www.mql5.com/en/docs/constants/chartconstants/enum_timeframes}
}
\seealso{
\code{\link{MT5.DrawHorizontalLine()}}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
