% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.ServerTime}
\alias{MT5.ServerTime}
\title{MT5 server time}
\usage{
MT5.ServerTime()
}
\value{
Returns \code{POSIXct} date-time of MT5 server time. It may \strong{not} be local machine time.
}
\description{
Obtain MT5 server time.
}
\examples{
\dontrun{

dtServerTime <- MT5.ServerTime()

## Extract hour
strftime(dtServerTime,format="\%H")

## Extract day
strftime(dtServerTime,format="\%d")

}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
