% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.ShowOrders}
\alias{MT5.ShowOrders}
\title{Show orders}
\usage{
MT5.ShowOrders()
}
\value{
Returns \emph{Data.frame} \eqn{[nx7]}, with follow informations:
\itemize{
  \item sSymbol \code{{character}}: order's symbol.
  \item {iType \code{{int}} order's type, for more see references:
         \itemize{
                 \item{2: ORDER_TYPE_BUY_LIMIT}
                 \item{3: ORDER_TYPE_SELL_LIMIT}
                 }
         }
  \item fVolume \code{{numeric}}: order's volume.
  \item fPrice \code{{numeric}}: order's price.
  \item fStop \code{{numeric}}: order's stop loss. For stop loss not set is returned \code{0}.
  \item fGain \code{{numeric}}: order's stop gain. For stop gain not set is returned \code{0}.
  \item iTicket \code{{int}}: order's ticket. Same displayed on the "Trade" tab. Used in other functions in \emph{mt5R} package.
}
}
\description{
Show all active orders displayed on the "Trade" tab. For active positions use \code{MT5.ShowPositions()} instead.
}
\examples{
\dontrun{

MT5.ShowOrders()

## Returns
##   sSymbol iType fVolume fPrice   fStop fGain iTicket
## 1  NZDUSD     2    0.10  0.682 0.00000     0  763153
## 2  NZDUSD     2    0.10  0.661 0.00000     0  763194
## 3  GBPUSD     2    0.01  1.330 1.32956     0  766489

}
}
\references{
\url{https://www.mql5.com/en/articles/2513}
}
\seealso{
\code{\link{MT5.ShowPositions()}}, \code{\link{MT5.DeleteOrder()}}, \code{\link{MT5.ModifyOrder()}}, \code{\link{MT5.MultipleOrders()}}, \code{\link{MT5.SingleOrder()}}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
