% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.ShowPositions}
\alias{MT5.ShowPositions}
\title{Show positions}
\usage{
MT5.ShowPositions()
}
\value{
Returns \emph{Data.frame} \eqn{[nx8]}, with follow informations:
\itemize{
\item sSymbol \code{{character}}: position's symbol.
\item {iCmd \code{{int}} position's type, for more see references:
\itemize{
\item{0: POSITION_TYPE_BUY}
\item{1: POSITION_TYPE_SELL}
}
}
\item fVolume \code{{numeric}}: position's volume.
\item fPrice \code{{numeric}}: position's price.
\item fStop \code{{numeric}}: position's stop loss. For stop loss not set is returned \code{0}.
\item fGain \code{{numeric}}: position's stop gain. For stop gain not set is returned \code{0}.
\item fProfit \code{{numeric}}: position's profit/loss. Same displayed on the "Trade" tab.
\item iTicket \code{{int}}: position's ticket. Same displayed on the MT5's "Trade" tab. Used in other functions in \emph{mt5R} package.
}
}
\description{
Show all active positions displayed on the "Trade" tab. For active orders use \code{MT5.ShowOrders()} instead.
}
\examples{
\dontrun{

MT5.ShowPositions()

## Returns
##    sSymbol iCmd fVolume   fPrice  fStop fGain fProfit iTicket
## 1  CADJPY    1    0.01 81.09000 0.0000     0    0.47  766491
## 2  CADJPY    0    0.01 81.11700 0.0000     0   -1.00  766490
## 3  EURUSD    0    0.01  1.22592 0.0000     0    0.77  766488
## 4  NZDUSD    0    0.10  0.71160 0.7106     0   29.50  763964

}
}
\references{
\url{https://www.mql5.com/en/docs/constants/tradingconstants/positionproperties}
}
\seealso{
\code{\link{MT5.ShowOrders()}}, \code{\link{MT5.ClosePosition()}}, \code{\link{MT5.ModifyPosition()}}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
