% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.SingleOrder}
\alias{MT5.SingleOrder}
\title{Send single order}
\usage{
MT5.SingleOrder(
  sSymbol,
  iCmd,
  fVol,
  fPrice = 0,
  fStop = 0,
  fGain = 0,
  iFillType = 0
)
}
\arguments{
\item{sSymbol}{character; symbol for trading.}

\item{iCmd}{int; operation type: \itemize{
\item{0: \code{ORDER_TYPE_BUY_LIMIT}}
\item{1: \code{ORDER_TYPE_SELL_LIMIT}}
}}

\item{fVol}{numeric; number of lots.}

\item{fPrice}{numeric; order price. \strong{WARNING}: if \code{fPrice = 0} it will be send as market order.}

\item{fStop}{numeric; if applicable use any value > \code{0}.}

\item{fGain}{numeric; if applicable use any value > \code{0}.}

\item{iFillType}{int; order type filling:
\itemize{
\item{0: \code{ORDER_TYPE_BUY_LIMIT} (default)}
\item{1: \code{ORDER_TYPE_SELL_LIMIT} (Market Order)}
\item{2: \code{ORDER_FILLING_IOC}}
}}
}
\value{
Returns \code{TRUE} if the order was successfully sent, \code{FALSE} otherwise.
}
\description{
Function used to open market or place a pending order. Only accept only order at time, for multiple orders consider to use \code{MT5.MultipleOrders()} instead.

The switch for pending order and market order is using \code{fPrice}. Any value higher than \code{0} is considered a pending order.

Some Forex brokers take \code{200ms} to accept orders into market, so take this time on mind. Always check MT5's "Experts" tab for more information.
}
\details{
Not all columns need to be filled as \code{MT5.MultipleOrders()} request. See examples.

For \code{fPrice = 0}, a market order will be sent and it will automatically set \code{iFillType} to \code{1}.

For more information about \code{ORDER_TYPE_*} see references.

\emph{WARNING! INVALID STOPS}: take note that buy orders are rejected when order's stop loss are higher than symbol's ask price, and vice versa for sell orders. This is a sanity check made by MT5, not by mt5R. Always check MT5's "Experts" tab for more details when an order is rejected.
}
\examples{
\dontrun{
## Sell market order. Using arguments
MT5.SingleOrder("EURUSD", iCmd = 1, fVol =  0.01)

## Sell market order, same above. Omitting arguments
MT5.SingleOrder("EURUSD", 1, 0.01)

## Buy market order, using stop. fStop need to be lower than symbol's ask price, otherwise the order will be rejected.
MT5.SingleOrder("EURUSD", 0, 0.01, fStop = 1.200) ## Last price: 1.2242 18-12-20. 1.200 < 1.2242 = OK

## Sell pending order
MT5.SingleOrder("EURUSD", iCmd = 1, fVol =  0.01, fPrice = 1.2334)

## Pending Order
MT5.SingleOrder("EURUSD", iCmd = 1, fVol =  0.01, fPrice = 1.18320) ## Sell Pending Order
MT5.SingleOrder("EURUSD", iCmd = 0, fVol =  0.01, fPrice = 1.18320, fStop = 1.18200) ## Buy Pending Order with stop
MT5.SingleOrder("EURUSD", iCmd = 1, fVol =  0.01, fPrice = 1.18720, fStop = 1.18790) ## Sell Pending Order with stop

}
}
\references{
\url{https://www.mql5.com/en/docs/constants/tradingconstants/orderproperties}

\url{https://www.investopedia.com/terms/m/marketorder.asp}
}
\seealso{
\code{\link{MT5.DeleteOrder()}}, \code{\link{MT5.ModifyOrder()}}, \code{\link{MT5.MultipleOrders()}}, \code{\link{MT5.ShowOrders()}}, \code{\link{MT5.ClosePosition()}}, \code{\link{MT5.ModifyPosition()}}, \code{\link{MT5.ShowPositions()}}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
