% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.SymbolExpiration}
\alias{MT5.SymbolExpiration}
\title{Expiration symbol date}
\usage{
MT5.SymbolExpiration(sSymbol)
}
\arguments{
\item{sSymbol}{character; target symbol.}
}
\value{
Returns \code{Date} if symbol has date of expiration, \code{NA} otherwise.
}
\description{
Obtain expiration date of target \code{sSymbol}.
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
