% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.SymbolInMarketwatch}
\alias{MT5.SymbolInMarketwatch}
\title{Check if symbol is in MT5's marketwatch}
\usage{
MT5.SymbolInMarketwatch(sSymbol)
}
\arguments{
\item{sSymbol}{character; target symbol.}
}
\value{
Returns \code{TRUE} if symbol is in MT5's marketwatch, \code{FALSE} otherwise.
}
\description{
Function to check if \code{sSymbol} is in MT5's marketwatch.
}
\seealso{
\code{\link[=MT5.MarketwatchAdd]{MT5.MarketwatchAdd()}}, \code{\link[=MT5.MarketwatchRemove]{MT5.MarketwatchRemove()}}, \code{\link[=MT5.Marketwatch]{MT5.Marketwatch()}}
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
