% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{MT5.zExample}
\alias{MT5.zExample}
\title{Exemple function}
\usage{
MT5.zExample()
}
\description{
Example function to help other to create their own functions.

Note that we use "!" to subdivide the request.
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
