% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Source_MT5.R
\name{SUP_MT5_StackRequests}
\alias{SUP_MT5_StackRequests}
\title{Support function of mt5R package}
\usage{
SUP_MT5_StackRequests(sVetor, sOrdem)
}
\description{
This function is used in others mt5R functions to concatenate strings and save space, so they look more elegant! Yay

Into further updates this function probably will be retired.
}
\author{
Guilherme Kinzel, \email{guikinzel@gmail.com}
}
