% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt5reti.R
\name{symbol_select}
\alias{symbol_select}
\title{symbol_select}
\usage{
symbol_select(symbol, enable)
}
\arguments{
\item{symbol}{character}

\item{enable}{bool}
}
\value{
\itemize{
\item 
}
}
\description{
symbol_select
}
